@@ -2,7 +2,7 @@ package beta.qlife.ui.fragments;
 
 import android.app.Activity;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -25,7 +25,7 @@ import java.util.HashMap;
  * Created by Carson on 18/07/2017.
  * Fragment that displays the cafeterias in the phone database.
  */
-public class CafeteriasFragment extends android.support.v4.app.ListFragment implements ActionbarFragment, DrawerItem, ListFragment {
+public class CafeteriasFragment extends androidx.fragment.app.ListFragment implements ActionbarFragment, DrawerItem, ListFragment {
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
